+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 * DataTable: select a row when user focuses on an input in that row (e.g. via TAB).
 * DataTable: add `isItemSelected` to the props received by all column render functions.
+* DataTable: add `onRowDoubleClick` prop
 * New HeightMeasurer component: measures the height of its parent and passes it down to its function-as-a-child. **HeightMeasurer MUST BE the direct child of an element with an "extrinsic" height (ie. a height that is not determined by its children, but rather by its parents, e.g. a flex item with `overflow: hidden`)**
 * Add `simplifyString()` to public API (useful for simplistic sorting and filtering).
 * Bugfix: DataTable: Discard bubbled keyDown events when focus is on an input.